Add an ElementEditor decorator
This allows us to move helper classes into this class and
render the element editor partial in a more efficient way.
